# contracts-gen/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeStatusRequest": {
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "CreateExperimentRequest": {
        "properties": {
          "conversionGoal": {
            "nullable": true,
            "type": "string"
          },
          "flagKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            },
            "type": "array"
          }
        },
        "required": [
          "flagKey",
          "name",
          "variants",
          "conversionGoal"
        ],
        "type": "object"
      },
...
      "ExperimentResponse": {
        "properties": {
          "conversionGoal": {
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "flagKey": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            },
            "type": "array"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "flagKey",
          "name",
          "status",
          "conversionGoal",
          "version",
          "variants",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateExperimentRequest": {
        "properties": {
          "variants": {
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            },
            "type": "array"
          }
        },
        "required": [
          "variants"
        ],
        "type": "object"
      },
...
      "VariantDto": {
        "properties": {
          "key": {
            "type": "string"
          },
          "weight": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "weight"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OptiPulse.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/v1/experiments": {
      "get": {
        "operationId": "ListExperiments",
        "parameters": [
          {
            "in": "query",
            "name": "flagKey",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExperimentResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Experiments"
        ]
      },
      "post": {
        "operationId": "CreateExperiment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            },
            "description": "Created"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "tags": [
          "Experiments"
        ]
      }
    },
    "/api/v1/experiments/{id}": {
      "get": {
        "operationId": "GetExperiment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Experiments"
        ]
      },
      "put": {
        "operationId": "UpdateExperiment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "tags": [
          "Experiments"
        ]
      }
    },
    "/api/v1/experiments/{id}/status": {
      "post": {
        "operationId": "ChangeExperimentStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "tags": [
          "Experiments"
        ]
      }
    },
...
  },
...
}